Organize files to make a clean directory
