+++ schema/schema.sql
@@ -48,9 +48,9 @@ CREATE TABLE books (
   id SERIAL PRIMARY KEY,
   publisher VARCHAR,
   cover_state VARCHAR,
-  genre_id INT NULL REFERENCES genres(id) ON DELETE CASCADE,
-  author_id INT NULL REFERENCES authors(id) ON DELETE CASCADE,
-  label_id INT NULL REFERENCES labels(id) ON DELETE CASCADE,
+  genre_id INT NULL REFERENCES genres(id),
+  author_id INT NULL REFERENCES authors(id),
+  label_id INT NULL REFERENCES labels(id),
   publish_date DATE NOT NULL,
   archived BOOLEAN NOT NULL
 );
