+++ src/book_methods.rb
