+++ src/gaming_methods.rb
