+++ src/item.rb
@@ -40,6 +40,6 @@ class Item
   private
 
   def can_be_archived?
-    (Date.today - Date.parse(@publish_date)).to_i > 365 * 10
+    (Date.today.year - Date.parse(@publish_date).year) > 10
   end
 end
