+++ storage/book.json
@@ -1,1 +1,1 @@
-[{"ref":"#<Book:0x00000229323caf20>","value":{"id":77,"genre":null,"author":null,"label":null,"publish_date":null,"archived":false,"publisher":"Edison","cover_state":"goof"}}]
+[{"ref":"#<Book:0x000002900c6de018>","value":{"id":63,"genre":null,"author":null,"label":null,"publish_date":null,"archived":false,"publisher":"Edison","cover_state":"good"}},{"ref":"#<Book:0x000002900c6cf608>","value":{"id":67,"genre":null,"author":null,"label":null,"publish_date":null,"archived":false,"publisher":"Kai","cover_state":"bad"}}]
