+++ storage/game.json
@@ -1,1 +1,1 @@
-[{"ref":"#<Game:0x00000229323cb060>","value":{"id":902,"genre":null,"author":null,"label":null,"publish_date":"2021/01/25","archived":false,"multiplayer":"y","last_played_at":"2023/01/23"}},{"ref":"#<Game:0x00000229323cafc0>","value":{"id":999,"genre":null,"author":null,"label":null,"publish_date":"2019/09/09","archived":false,"multiplayer":"n","last_played_at":"2022/01/11"}}]
+[{"ref":"#<Game:0x000002900c905828>","value":{"id":570,"genre":null,"author":null,"label":null,"publish_date":"2021/01/25","archived":false,"multiplayer":"y","last_played_at":"2023/01/23"}},{"ref":"#<Game:0x000002900c905080>","value":{"id":744,"genre":null,"author":null,"label":null,"publish_date":"2019/09/09","archived":false,"multiplayer":"n","last_played_at":"2022/01/11"}}]
