+++ storage/music.json
@@ -1,1 +1,1 @@
-[{"ref":"#<MusicAlbum:0x00000229323cb920>","value":{"id":27,"genre":null,"author":null,"label":null,"publish_date":"02/01/2023","archived":false,"on_spotify":false}},{"ref":"#<MusicAlbum:0x00000229323cb880>","value":{"id":56,"genre":null,"author":null,"label":null,"publish_date":"12/01/2023","archived":false,"on_spotify":false}},{"ref":"#<MusicAlbum:0x00000229323cb7e0>","value":{"id":13,"genre":null,"author":null,"label":null,"publish_date":"02/01/2022","archived":false,"on_spotify":false}},{"ref":"#<MusicAlbum:0x00000229323c9d00>","value":{"id":64,"genre":null,"author":null,"label":null,"publish_date":"20/10/2019","archived":true,"on_spotify":true}}]
+[{"ref":"#<MusicAlbum:0x000002900c907290>","value":{"id":65,"genre":null,"author":null,"label":null,"publish_date":"02/01/2023","archived":false,"on_spotify":false}},{"ref":"#<MusicAlbum:0x000002900c906a98>","value":{"id":12,"genre":null,"author":null,"label":null,"publish_date":"12/01/2023","archived":false,"on_spotify":false}},{"ref":"#<MusicAlbum:0x000002900c906778>","value":{"id":94,"genre":null,"author":null,"label":null,"publish_date":"02/01/2022","archived":false,"on_spotify":false}},{"ref":"#<MusicAlbum:0x000002900c906458>","value":{"id":57,"genre":null,"author":null,"label":null,"publish_date":"20/10/2019","archived":false,"on_spotify":false}}]
